Moved hooks to their own files
